third_party/WebKit/LayoutTests/http/tests/images/image-currentsrc.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test.js"></script>
<body>
<script>
    document.body.innerHTML += '<img id="load" src="http://127.0.0.1:8000/resources/slow-image.php?name=square100.png&mimeType=image/png&sleep=100">';
    document.body.innerHTML += '<img id="error" srcset="http://127.0.0.1:8000/resources/slow-image.php?name=not_there.png&mimeType=image/png&sleep=300">';
    document.body.innerHTML += '<img id="redirect" src="http://127.0.0.1:8000/resources/slow-image.php?redirect_name=square100.png&mimeType=image/png&sleep=500">';
    jsTestIsAsync = true;
    var img = window.document.getElementById("load");
    shouldBe('img.currentSrc', '""');
    var expectedCurrentSrc = "http://127.0.0.1:8000/resources/slow-image.php?name=square100.png&mimeType=image/png&sleep=100";
    img.onload = function() {
        img.srcset = "http://127.0.0.1:8000/resources/slow-image.php?name=square200.png&mimeType=image/png&sleep=100 1x";
        shouldBe('img.currentSrc', 'expectedCurrentSrc');
        expectedCurrentSrc = "http://127.0.0.1:8000/resources/slow-image.php?name=square200.png&mimeType=image/png&sleep=100";
    };
    var img_err = window.document.getElementById("error");
    img_err.onerror = function() {
        shouldBe('img_err.currentSrc', '"http://127.0.0.1:8000/resources/slow-image.php?name=not_there.png&mimeType=image/png&sleep=300"');
    };

    var img_redirect = window.document.getElementById("redirect");
    img_redirect.onload = function() {
        shouldBe('img_redirect.currentSrc', '"http://127.0.0.1:8000/resources/slow-image.php?redirect_name=square100.png&mimeType=image/png&sleep=500"');
    };
    window.onload = function() {
        finishJSTest();
    };
</script>
</body>
